requirements.txt
faiss-cpu
torch
transformers
accelerate
fastapi
requests
easyocr
uvicorn
pydantic